RAG/requirements.txt
# Core RAG and AI libraries
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1

# Document processing
PyPDF2>=3.0.0
markdown>=3.4.0
python-docx>=0.8.11
openpyxl>=3.1.0

# OCR and Image Processing
pytesseract>=0.3.10
Pillow>=9.0.0
pdf2image>=1.16.0
opencv-python>=4.8.0

# Embeddings and vector storage
sentence-transformers>=2.2.0
chromadb>=0.4.0
faiss-cpu>=1.7.0

# UI
streamlit>=1.28.0

# Ollama integration
ollama>=0.1.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
